Public dashboard

+++ app/api/admin/offices/route.ts
@@ -1,9 +1,18 @@
+import dbconnect from "@/lib/dbConnect";
+import { checkAdmins } from "@/lib/userAuth";
 import { OfficeModel } from "@/models/OfficeModel";
 import ShowError from "@/utils/ShowError";
 import { NextRequest, NextResponse } from "next/server";
 
+
 export async function GET(req: NextRequest) {
     try {
+        await dbconnect();
+        // const loggedUser = await checkAdmins(req);
+        // if(!loggedUser || loggedUser.role !== "superadmin"){
+        //     return ShowError(401, "Unauthorized. Login Again.");
+        // }
+        
         const officeList = await OfficeModel.find();
         return NextResponse.json(officeList,{ status: 200 });
     } catch (error: any) {
@@ -13,12 +22,23 @@ export async function GET(req: NextRequest) {
 
 export async function POST(req: NextRequest) {
     try {
+        await dbconnect();
+        const loggedUser = await checkAdmins(req);
+        if(!loggedUser || loggedUser.role !== "superadmin"){
+            return ShowError(401, "Unauthorized. Login Again.");
+        }
         const { name, province, address, district } = await req.json();
         if (!name || !province || !address || !district) {
             return ShowError(400, "Invalid request. Missing fields");
         }
-        const office = await OfficeModel.create({ name, province, address, district });
-        return NextResponse.json(office, { status: 200 });
+        const exists = await OfficeModel.findOne({
+            name
+        })
+        if (exists) {
+            return ShowError(400, "Office already exists");
+        }
+        const office = await OfficeModel.create({ name, province, address, district, createdBy: loggedUser._id});
+        return NextResponse.json(office, { status: 201 });
     } catch (error: any) {
         return ShowError(500, error.message);
     }
@@ -26,6 +46,11 @@ export async function POST(req: NextRequest) {
 
 export async function PUT(req: NextRequest) {
     try {
+        await dbconnect();
+        const loggedUser = await checkAdmins(req);
+        if(!loggedUser || loggedUser.role !== "superadmin"){
+            return ShowError(401, "Unauthorized. Login Again.");
+        }
         const { id, content } = await req.json();
         const office = await OfficeModel.findByIdAndUpdate(id, content, { new: true });
         if (!office) {
@@ -40,6 +65,11 @@ export async function PUT(req: NextRequest) {
 
 export async function DELETE(req: NextRequest) {
     try {
+        await dbconnect();
+        const loggedUser = await checkAdmins(req);
+        if(!loggedUser || loggedUser.role !== "superadmin"){
+            return ShowError(401, "Unauthorized. Login Again.");
+        }
         const id = req.nextUrl.searchParams.get("id");
         if(!id) return ShowError(400, "Invalid request. Missing id");
         const office = await OfficeModel.findByIdAndDelete(id);
